pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "side-builder"
version = "0.1.9"
authors = [ { name = "Soonki Ji", email = "soonki.ji@gmail.com" } ]
description = "AI-powered Selenium IDE (.side) player and builder for Selenium and Playwright"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "click>=8.3.1",
    "openai>=2.24.0",
    "playwright>=1.58.0",
    "python-dotenv>=1.2.2",
    "selenium>=4.41.0",
]
classifiers = [
    "Programming Language :: Python :: 3.10",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.scripts]
side-builder = "side_player.builder:main"

[project.urls]
Homepage = "https://github.com/tayaee/side-builder"
Repository = "https://github.com/tayaee/side-builder.git"
Issues = "https://github.com/tayaee/side-builder/issues"

[dependency-groups]
dev = []

[tool.hatch.build.targets.wheel]
packages = ["src/side_player"]
